+++ paddle/fluid/train/custom_trainer/feed/process/learner_process.h
@@ -21,9 +21,9 @@ public:
 
 protected:
 //同步保存所有模型
-virtual int wait_save_model(int epoch_id, ModelSaveWay way);
+virtual int wait_save_model(uint64_t epoch_id, ModelSaveWay way);
 //异步保存指定模型
-virtual std::future<int> save_model(int epoch_id, int table_id, ModelSaveWay way);
+virtual std::future<int> save_model(uint64_t epoch_id, int table_id, ModelSaveWay way);
 //执行指定训练网络
 virtual int run_executor(Executor* executor);
